Condition Message Finder Integration Code
