+++ core/CMIF/utils/instrumentation.py
@@ -1,0 +1,25 @@
+import json
+
+
+def dump_jsonl(data, output_path, append=False):
+    """
+    Write list of objects to a JSON lines file.
+    """
+    mode = 'a+' if append else 'w'
+    with open(output_path, mode, encoding='utf-8') as f:
+        for line in data:
+            json_record = json.dumps(line, ensure_ascii=False)
+            f.write(json_record + '\n')
+    print('Wrote {} records to {}'.format(len(data), output_path))
+
+
+def load_jsonl(input_path):
+    """
+    Read list of objects from a JSON lines file.
+    """
+    data = []
+    with open(input_path, 'r', encoding='utf-8') as f:
+        for line in f:
+            data.append(json.loads(line.rstrip('\n|\r')))
+    print('Loaded {} records from {}'.format(len(data), input_path))
+    return data
