+++ main.py
@@ -1,14 +1,21 @@
 # Import your method from core and pass user data
 
 import argparse
+
+from test.CMIF.testing import predict_code_t5
+
+
 # add the imports for your methods
 
 def get_method(method_name):
     # We need to validate here if the input method_name exist in our method or not
     # check your method name
-    if method_name.lower() == 'method_name':
-        # return your method class 
-        return YOUR_METHOD()
+    # if method_name.lower() == 'method_name':
+    #     return your method class
+
+    if method_name.lower() == 'cmif':
+        # return your method class
+        return predict_code_t5
     
     # Add more checkers as needed
     else:
@@ -28,6 +35,7 @@ def main():
     checker = get_method(args.method)
     
     #call the method object and pass user data
+    checker(args.input_dir, args.output_dir)
 
 if __name__ == "__main__":
     main()
